Change default template and add Bootstrap-Flask dependency
Modified the default route of the application to render a game library index instead of an individual game description. This improvement enhances user navigation by providing an overview of available games on the home page.

Added 'Bootstrap-Flask' to the requirements.txt to enable efficient rendering of HTML templates, thereby enhancing app UX and performance.