Update gameLibrary view to render selected genre games
A new parameter, slide_genre_games, has been added to the render_template function in the gameLibrary.py file. The parameter is assigned a subsection of the selected_genre_games list. This was done to feature a subset of games from the selected genre in the slide view, enhancing content delivery and viewer engagement.