Refactored game search functionality, moved to separate blueprint
The game search functionality using criteria query was moved from gameLibrary to its own seperate homePage blueprint. This change came from a need to make the application more modular and improve maintainability. Game model was also updated to store categories and tags in sets rather than lists for better efficiency. Files related to the gameLibrary was removed as it became obsolete with the new design."