Adjust sidebar, header style for better User Interface
To improve website's visual appeal, made changes in the 'sidebar.css' and 'header.css'. The background-color for the sidebar changed to a darker shade (#041C32) for better contrast with text. Likewise, hover color for sidebar links changed to #d100ff for more visibility. In 'header.css', the header's background-color aligned with the sidebar (#041C32) for consistency across UI. The z-index of the hamburger menu increased to 9999 to remain above all elements. Similarly, the hover color of header links modified to #c600d0 for better UI experience.