Add new footer, update homepage layout and style
This commit adds a new footer to the site layout by creating a new HTML template and corresponding CSS file. The index page has been updated to include this footer and the look of the homepage has been refreshed overall to provide a more welcoming and professional aesthetic. This includes the addition of an 'about us' section and stylistic changes to improve alignment and readability. These changes aim to improve the user experience and make the site feel more cohesive.