"Adjusted 'searchResults' layout and styles"
Moved the include for 'footer.html' to below the body tag in searchResults.html, thus ensuring a consistent footer across all pages.

In searchResults.css, I added styles to set the height of html and body to 100% and set body to be a flex container which allow the main content to expand to take available space ensuring footer will be at bottom always. I also improved the styling of the individual game display units within the search results for better readability and user experience.