Update CSS for better readability and maintainability. Made header stick to top of page on scroll.
The CSS files for game library have been refactored with a focus on improving the readability of the code and increasing the maintainability of the stylesheets.

The CSS has been standardized across all style sheets with addition of CSS variables, clearer class definitions, better indentation and spacing, and removal of redundant or commented-out code. Critical sections like Navigation, Header, and Footer have been clearly annotated.

These changes should make it easier for developers to understand the styling of the website, alter existing styles or add new ones, thereby improving productivity and decreasing the likelihood of creating bugs.

Finally, the use of CSS variables will improve consistency of the design across the site and make thematic changes easier to implement. The changes do not affect the visual appearance of the site.