Uncomment code to raise exception for an unknown user
In the 'games/authentication' file, uncommented two lines of code to activate an exception when an unknown user tries to authenticate. This change was necessary to enhance user security by preventing non-existing users from attempting to gain access to the system. Now, an UnknownUserException will be raised if such a scenario occurs.