Update test values in domainmodel tests
Updated the expected return values in few test cases of the domainmodel test suite as the data has grown in size (HAHAHA). The data size determines these values concerning games, publishers, and genres. Also, one of the publishers name has been updated to reflect the current dataset. These changes were necessary to sync the test cases with the current dataset and for the tests to pass successfully.