Improve header UI and add user avatar
Changes were made to enhance the user interface of the header. The search bar and 'login' and 'register' / 'user profile' and 'logout' options are now displayed in one line, making it more user-friendly and aesthetically pleasing. A user avatar, represented by an SVG, is now present beside the username when a user is logged in - adding a more personalized and visually appealing touch. The styles defined in the CSS file were adjusted accordingly to align with these changes. Also, some HTML formatting was altered for better readability. This enhances user experience and makes the code easier to understand and maintain.