Replace hard-coded margin in header with CSS variable
In order to improve maintainability of the code, a hard-coded margin-left value in .navigation class of header.css file was replaced with a CSS variable --NAV-MARGIN. Now, instead of changing the margin value in each class where it's applied, we can adjust it in one place, thereby improving the efficiency of style changes. Also, instances of unused code i.e., color attribute for nav__user_logo svg, were deleted for code clean-up.