Remove auth.css and update styles for consistency
Removed the obsolete 'auth.css' file as the authentication pages now utilize the newly defined styles in the 'userProfile.css'. This change ensures consistency across all pages and contributes to a smoother user experience.

Added a set of base colors in 'userProfile.css' for easy references and modifications in the future. This promotes better color management and aids maintainability.

Updated 'userProfile.css', 'gameLibStyle.css', and 'index.css' files to make use of the new base colors. This implements the changes across the site and achieves a consistent color scheme.