Improve game container alignment and readability
Updated 'gameLibStyle.css' and 'gameLibrary.html' to enhance aligning in the game container and improve the readability of various elements. The CSS change enables the forms within game info to align at the start of their flex container, ensuring consistency throughout the website.

In the 'gameLibrary.html' file, made changes to indentation to improve readability and maintenance of the code. Moreover, the accessibility of the 'View Game' link in the game box has been improved. It's surrounded by an 'a' tag, making the entire box clickable, enhancing the user experience.