Refactor variable naming and integrate SQLAlchemy
In several functions across multiple modules, the `id` and `title` parameter names were renamed to `game_id` and `game_title` to improve code readability and understanding. Introduced SQLAlchemy by adding it to the requirements.txt file and implemented it in the new database_repository module. This change was made to leverage SQLAlchemy's ORM capabilities for easier database management. Added documentation to Game class.