Change default value in create_app function
The default value of the 'test_config' in the 'create_app' function in games/__init__.py file was changed from 'True' to 'None'. This adjustment was implemented because it reflects more accurately the behavior of the function when no config is passed and leads to less confusion while handling configurations of the Flask application.